+++ WakaTime/Metadata.cs
@@ -1,0 +1,10 @@
+﻿namespace WakaTime
+{
+    public class Metadata
+    {
+        public string EditorVersion { get; set; }
+        public string EditorName { get; set; }
+        public string PluginVersion { get; set; }
+        public string PluginName { get; set; }
+    }
+}