+++ HISTORY.rst
@@ -3,6 +3,17 @@ History
 -------
 
 
+5.0.0 (2002-12-04)
+++++++++++++++++++
+
+- Fix error "This ANSI Plugin is not compatible with your Unicode Notepadd++".
+  `#46 <https://github.com/wakatime/notepadpp-wakatime/issues/46>`
+- Update to .net frmework 4.8.
+- Add lines in file and current line number.
+- Update `PluginInfrastructure` to latest version.
+- Remove dependency do nuget library.
+
+
 4.3.0 (2022-06-05)
 ++++++++++++++++++